+++ MailSender/Tests/MailSender.ConsoleTest/MailSender.ConsoleTest.csproj
@@ -45,6 +45,7 @@
   <ItemGroup>
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="ThreadTest.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
